Do not return iface with only a Link-local address
This was done because it was noticed that if a network interface has the
wait_for_lease argument set to true, then it could proceed having
received solely a Link-local address. When other Terraform resources
refer to this address to connect to it, then the resources will more
than likely never finish completing, as Link-local addresses require a
specific setup on the host.